refactor(transcribe): receive a general readable stream as input
