+++ packages/service/core/ai/audio/transcriptions.ts
@@ -1,4 +1,4 @@
-import type fs from 'fs';
+import type { Readable } from 'node:stream';
 import { getAxiosConfig } from '../config';
 import { axiosWithoutSSRF } from '../../../common/api/axios';
 import FormData from 'form-data';
@@ -8,10 +8,12 @@ import { UserError } from '@fastgpt/global/common/error/utils';
 export const aiTranscriptions = async ({
   model: modelData,
   fileStream,
+  filename,
   headers
 }: {
   model: STTModelType;
-  fileStream: fs.ReadStream;
+  fileStream: Readable;
+  filename: string;
   headers?: Record<string, string>;
 }) => {
   if (!modelData) {
@@ -20,7 +22,7 @@ export const aiTranscriptions = async ({
 
   const data = new FormData();
   data.append('model', modelData.model);
-  data.append('file', fileStream);
+  data.append('file', fileStream, { filename });
 
   const aiAxiosConfig = getAxiosConfig();
