+++ projects/app/src/pages/api/v1/audio/transcriptions.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { jsonRes } from '@fastgpt/service/common/response';
-import { pushWhisperUsage } from '@/service/support/wallet/usage/push';
+import { pushWhisperUsage } from '@fastgpt/service/support/wallet/usage/controller';
 import { authChatTargetCrud } from '@/service/support/permission/auth/chat';
 import { NextAPI } from '@/service/middleware/entry';
 import { aiTranscriptions } from '@fastgpt/service/core/ai/audio/transcriptions';
@@ -13,6 +13,7 @@ import { multer } from '@fastgpt/service/common/file/multer';
 import { AudioTranscriptionsDataSchema } from '@fastgpt/global/openapi/core/chat/record/api';
 import { parseApiInput } from '@fastgpt/service/common/zod/requestParseError';
 import { ERROR_ENUM } from '@fastgpt/global/common/error/errorCode';
+import { UsageSourceEnum } from '@fastgpt/global/support/wallet/usage/constants';
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
   const filepaths: string[] = [];
@@ -58,13 +59,15 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
 
     const transcriptionsResult = await aiTranscriptions({
       model: getDefaultSTTModel(),
-      fileStream: result.getReadStream()
+      fileStream: result.getReadStream(),
+      filename: result.fileMetadata.originalname
     });
 
     pushWhisperUsage({
       teamId,
       tmbId,
-      duration: transcriptionsResult?.usage?.total_tokens || duration
+      duration: transcriptionsResult?.usage?.total_tokens || duration,
+      source: UsageSourceEnum.fastgpt
     });
 
     jsonRes(res, {
